Refactor the tests to be one endpoint per test file
