+++ apps/back-end/tests/server/auth/helpers/COOKIES.ts
@@ -1,0 +1,7 @@
+const COOKIES = [
+  "oauth_state=valid-state",
+  "oauth_pkce_verifier=test-verifier",
+  "oauth_redirect=http://localhost:5173",
+];
+
+export default COOKIES;
